excluí tabela ubs, troquei tabela hospital por centro-medico, add tipo rm centro_medico

+++ src/controllers/users.js
@@ -14,6 +14,17 @@ exports.createUsers = async (req, res) => {
     id_endereco,
   } = req.body;
 
+  const errors = []
+
+  if(!nome) {
+    errors.push({error: "Nome is empty"})
+  }
+
+  if(!cpf) {
+    errors.push({error: "CPF is empty"})
+  }
+
+
   const cpfUnico = 'SELECT * FROM users WHERE cpf = $1';
   db.query(cpfUnico, [cpf]).then(function(resultadoDaConsultaCpf) {
     if (resultadoDaConsultaCpf.rows.length > 0) {
@@ -21,8 +32,11 @@ exports.createUsers = async (req, res) => {
     }
   });
 
+  if (errors.length > 0) {
+    return res.status(400).json(errors)}
+
   const {rows} = await db.query(
-    'INSERT INTO users (nome, data_nascimento, telefone, cpf, numero_cartao_sus, celular, email, id_endereco) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)',
+    'INSERT INTO users (nome, data_nascimento, telefone, cpf, numero_cartao_sus, celular, email, id_endereco, id_role) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)',
     [
       nome,
       data_nascimento,
@@ -45,9 +59,12 @@ exports.createUsers = async (req, res) => {
       celular,
       email,
       id_endereco,
+      id_role,
     },
   });
 };
+
+
 // ==> Método responsável por listar todos os 'Users':
 exports.listAllUsers = async (req, res) => {
   const response = await db.query('SELECT * FROM users ORDER BY nome ASC ');
@@ -71,10 +88,11 @@ exports.updateUserById = async (req, res) => {
     celular,
     email,
     id_endereco,
+    id_role,
   } = req.body;
 
   const response = await db.query(
-    'UPDATE users SET nome = $1, data_nascimento = $2, telefone = $3, numero_cartao_sus = $4, celular = $5, email = $6, id_endereco = $7 WHERE cpf = $8',
+    'UPDATE users SET nome = $1, data_nascimento = $2, telefone = $3, numero_cartao_sus = $4, celular = $5, email = $6, id_endereco = $7, id_role = $8 WHERE cpf = $9',
     [
       nome,
       data_nascimento,
@@ -84,6 +102,7 @@ exports.updateUserById = async (req, res) => {
       email,
       id_endereco,
       cpf,
+      id_role,
     ],
   );
 
@@ -93,9 +112,7 @@ exports.updateUserById = async (req, res) => {
 // ==> Método responsável por excluir um 'User' pelo 'Id':
 exports.deleteUserById = async (req, res) => {
   const userId = parseInt(req.params.id);
-  await db.query('DELETE FROM users WHERE id = $1', [
-    userId
-  ]);
+  await db.query('DELETE FROM users WHERE id = $1', [userId]);
 
-  res.status(200).send({ message: 'User deleted successfully!', userId });
+  res.status(200).send({message: 'User deleted successfully!', userId});
 };
